Add files via upload
These files include the initial proposal and visualizations on what this build entails.  I did a lot of research and took a stab at outlining this project based on that research and the course material.  I may add additional docs in the future as my learning shapes up.  I wanted to leave a record of where I started.