Add files via upload
I deleted the original loose files, not sure if the documentation I wrote also got deleted.  I re-uploaded in a folder, I want to keep this repository as organized as I can.  
This folder includes the initial proposal in a template format provided by Springboard.  In addition, I did research on methods on how to go about planning a website.  I combined the research and the course material to take a first stab at visualizing how this build will go.  I wanted to document where I started on this project because things will surely change as my learning takes shape.